Remember to use double's in DynamicPriceProvider

+++ src/com/pablo67340/guishop/definition/DynamicPriceProvider.java
@@ -2,9 +2,9 @@ package com.pablo67340.guishop.definition;
 
 public interface DynamicPriceProvider {
 	
-	double calculateBuyPrice(String item, int quantity, int staticBuyPrice, int staticSellPrice);
+	double calculateBuyPrice(String item, int quantity, double staticBuyPrice, double staticSellPrice);
 	
-	double calculateSellPrice(String item, int quantity, int staticBuyPrice, int staticSellPrice);
+	double calculateSellPrice(String item, int quantity, double staticBuyPrice, double staticSellPrice);
 	
 	void buyItem(int quantity);
